MICROSERVICE/requirements.txt
django-2.2.2
djangorestframework-3.9.4
jinja2-2.10.1
xhtml2pdf-0.2.3
firebase-3.0.1